[PECOFF] Fix bug that section grouping may lay out atoms in a wrong order.
We added layout edges to the head atoms in grouped sections. That was wrong,
because the head atom needs to be followed by the other atoms in the *same*
section, not by the other section contents. With this patch, layout edges are
added from tail atom, which is the last atom in a section, to head atom.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@189573 91177308-0d34-0410-b5e6-96231b3b80d8