[Darwin] Fix Reference to nullptr.
On discussing this with Nick, it looks like the StubAtoms
that contain a lazyImmediate reference kind should be null
and the location needs to be fixed up later with some value
that is an offset into the __LINKEDIT segment.

The drawback is that it allows yaml files with references
that expect a target to be considered without one.

This results in bad yaml files that would need to be handled
in the YAML Reader.

Inorder to fix this, the Stub Atoms use a dummy target such
as itself.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@193476 91177308-0d34-0410-b5e6-96231b3b80d8