[PECOFF] Use INT3 instead of NOP.
This patch does not change the meaning of the program, but if something's wrong
in the linker or the compiler and the control reaches to the gap of imported
function table, it will stop immediately because of the presence of INT3. If
NOP, it'd fall through to the next call instruction, which is usually a
completely foreign function call.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@194860 91177308-0d34-0410-b5e6-96231b3b80d8