[PECOFF] Set OrdinalBase field in the export table.
OrdinalBase is an addend to the ordinals. We used to always set 1 to the field.
Although it produced a valid a DLL export table, it'd be a waste if the first
ordinal does not start with 1 -- we had to have NULL fields at the beginning of
the export address table. By setting the ordinal base, we can eliminate the
NULL fields.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@197367 91177308-0d34-0410-b5e6-96231b3b80d8