[PECOFF] Warn only when /export options are not compatible.
Currently LLD always print a warning message if the same symbol is specified
more than once for /export option. It's a bit annoying because specifying the
same symbol with compatible options is actually safe and considered as a
normal use case. This patch makes LLD to warn only when incompatible export
options are given.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@198104 91177308-0d34-0410-b5e6-96231b3b80d8