[PECOFF] Add a test for lib.exe subcommand.
Previously the input file for the lib.exe command would be removed
as soon as the command exits, so we couldn't write a test to check
the file contents are correct.

This patch adds /lldmoduledeffile: option to retain a copy of the
temporary file at the given file path, so that you can see the file
if you want.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@207727 91177308-0d34-0410-b5e6-96231b3b80d8