[PECOFF] Remember /noentry option so that later passes can handle it.
This is a part of a larger change to move the entry point
processing to a later pass than the driver. On Windows the default
entry point function varies depending on user-provided functions.
That means the driver is not able to correctly know the entry point
function name. Only passes after the core linker can infer it.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@213697 91177308-0d34-0410-b5e6-96231b3b80d8