[PECOFF] Fix PE+ relocations
The implementation of AMD64 relocations was imcomplete
and wrong. On AMD64, we of course have to use AMD64
relocations instead of i386 ones. This patch fixes the
issue.

LLD is now able to link hello64.obj (created from
hello64.asm) against user32.lib and kernel32.lib to
create a Win64 binary.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@216253 91177308-0d34-0410-b5e6-96231b3b80d8