[mach-o] Add support for -dependency_info command line option
This option is added by Xcode when it runs the linker.  It produces a binary
file which contains the file the linker used.  Xcode uses the info to
dynamically update it dependency tracking.

To check the content of the binary file, the test case uses a python script
to dump the binary file as text which FileCheck can check.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@219039 91177308-0d34-0410-b5e6-96231b3b80d8