Core: switch error mapping to switches
Rather than a series of cascading ifs, use a switch statement to convert the
error code to a string.  This has the benefit of allowing the compiler to inform
us if we ever add a new error code but fail to update the string representation.

Add in stringified versions for a couple of missing InputGraphErrors.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@219089 91177308-0d34-0410-b5e6-96231b3b80d8