[ELF] Add --as-needed.
The previous default behavior of LLD is --as-needed. LLD linked
against a DSO only if the DSO file was actually used to link an
executable (i.e. at least one symbol was resolved using the shared
library file.)

In this patch I added a boolean flag to FileNode for --as-needed.
I also added an accessor to DSO name to shared library file class.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@226274 91177308-0d34-0410-b5e6-96231b3b80d8