PE/COFF: rework how we handle base relocations
Generalise the base relocation handling slightly to support multiple base
relocation types in PE/COFF.  This is necessary to generate proper executables
for WoA.

Track the base relocation type from the decision that we need a base relocation
to the point where we emit the base relocation into base relocation directory.

Remove an outdated TODO item while in the area.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@226335 91177308-0d34-0410-b5e6-96231b3b80d8