PECOFF: Fix dummy symbol table in executable.
If the name field of a symbol table entry is all zero, it's interpreted
as it's pointing to the beginning of the string table. The first four
bytes of the string table is the size field, so dumpbin dumps that number
as an ASCIZ string.

This patch fills a dummy value to name field.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@228971 91177308-0d34-0410-b5e6-96231b3b80d8