PECOFF: Do not add extraneous symbols to the dead strip root.
Previously we added all undefined symbols found in object files to
the dead strip root. This patch makes the linker to stop doing that.
Undefined symbols would be resolved anyway, so this patch doesn't
change the linker behavior. It should slightly improve performance
but it's really marginal. This is a cleanup.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@231545 91177308-0d34-0410-b5e6-96231b3b80d8