Resolver: optimize fallback atoms.
Atoms with fallback atoms are never be added to the symbol table.
However, we added such atoms to _undefines array. We had to call
isCoalescedAway to identify and skip them. We should just stop
adding them in the first place.

This seems to make the linker ~1% faster in my test case.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@231552 91177308-0d34-0410-b5e6-96231b3b80d8