Resolver: Fix incorrect DenseMap mapping info.
Previously, getEmptyKey and getTombstoneKey return the same value
in the sense of isEqual defined by the same class, although they
need to be distinct values. This could confuse DenseMap.

We didn't see any issue by this wrong code because we don't delete
elements from the symbol table. We only add or replace elements.
But this is a bug and needs to be fixed anyway.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@231618 91177308-0d34-0410-b5e6-96231b3b80d8