[GNU/ELF] Add support for -X/--discard-locals.
There's (almost) never need to keep .L symbols around for production
builds. In fact, the FreeBSD kernel explicitly specify -X beacuse the
size impact (and the subsequent performance impact) might be significant,
because we keep symbols in memory.
I was tempted to make this the default, but I haven't (yet).

PR:		23232


git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@235357 91177308-0d34-0410-b5e6-96231b3b80d8