COFF: Cache raw pointers to relocation tables.
Getting an iterator to the relocation table is very hot operation
in the linker. We do that not only to apply relocations but also
to mark live sections and to do ICF.

libObject's interface is slow. By caching pointers to the first
relocation table entries makes the linker 6% faster to self-link.

We probably need to fix libObject as well.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@240603 91177308-0d34-0410-b5e6-96231b3b80d8