Centralize the handling of r_addend. NFC.
When a relocation points to a SHF_MERGE section, the addend has special meaning.
It should be used to find what in the section the relocation points to. It
should not be added to the output position.

Centralizing it means that the above rule will be implemented once, not once
per target.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@250421 91177308-0d34-0410-b5e6-96231b3b80d8