[lld][MachO] Always reserve space for the empty string in the mach-o symbol table.
The first entry in the MachO symbol table is always the empty string: make sure
we reserve space for it, or we will overflow the symbol table by one byte.

No test case - this manifests as an occasional memory error. In the near future
I hope to set up a bot building and runnnig LLD with sanitizers - that should
catch future instances of this issue.



git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@255178 91177308-0d34-0410-b5e6-96231b3b80d8