Set CMake ADDITIONAL_HEADER_DIRS on libraries with headers in /include.  NFC.
In a UI such as XCode, it can group the headers for a library with that library.
This is done in the CMakeLists.txt for the library itself by setting the path(s)
as ADDITIONAL_HEADER_DIRS.

LLVM already does this for all of its libraries, so just adding this to lld to
make things easier.  Should be NFC.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@257002 91177308-0d34-0410-b5e6-96231b3b80d8