Fix offset in test case.  NFC
It wasn't actually pointing to the function at the start of the text section, and so the offset in the binary differed when we passed the file through a second time.

The __eh_frame section uses implicit relocations and when reducing this test case from explicit to implicit, I got
the offset wrong.  This makes sure it is correct.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@257101 91177308-0d34-0410-b5e6-96231b3b80d8