Check for mismatched arch and OS when linking MachO files.
This patch makes use of the handleLoadedFile hook added in r257814.

That method is used to check the arch and the OS of the files we are linking
against the arch and OS on the context.

The first test to use this ensures that we do not try to combine i386 Mac OS code
with i386 simulator code.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@257837 91177308-0d34-0410-b5e6-96231b3b80d8