Avoid almost duplication in .dynamic finalize and write.
There are a few cases where we have almost duplicated code.

This patches fixes the simplest: the finalize and write of dynamic
section. Right now they have to have exactly the same structure to
decide if a DT_* entry is needed and then to actually write it.

We cannot just write it to a std::vector in the first pass since
addresses have not been computed yet.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@258723 91177308-0d34-0410-b5e6-96231b3b80d8