Use an ilist instead of std::list.  NFC.
The TrieNode/TrieEdge data structures here are allocated in a bumpptrallocator.

Unfortunately, TrieNode contained a std::list<TrieEdge> and as the allocator doesn't
call the TrieNode destructor, we ended up leaking the memory allocated by the std::list
itself.

Instead we can use an intrusive list as then we save the extra allocations anyway.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@258725 91177308-0d34-0410-b5e6-96231b3b80d8