ELF: Do not use fatal in LinkerScript.cpp.
This patch adds "Error" field to LinkerScript class. That field
is false by default, and set to true if there is a syntax error
in an input file. The linker script parser is a recursive-descedent
parser. Each function returns if Error is true -- so that
eventually the whole parser returns to a caller.

http://reviews.llvm.org/D16667

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@259557 91177308-0d34-0410-b5e6-96231b3b80d8