Default to an unknown OS instead of MacOSX.
Defaulting to unknown matches ld64, but it also makes sure that all
of our code can handle not knowing the platform.  For example, a later
commit will add support for version min load commands with an unknown
platform, which is a feature supported by ld64.

No test case here.  The next commit will have one with the version min
code that needed this patch.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@259739 91177308-0d34-0410-b5e6-96231b3b80d8