Fix more cases of UB from allocating 0 sized data.  NFC.
The size of a section can be zero, even when it contains atoms, so
long as all of the atoms are also size 0.  In this case we were
allocating space for a 0 sized buffer.

Changed this to only allocate when we need the space, but also cleaned
up all the code to use MutableArrayRef instead of uint8_t* so its much much
safer as we get bounds checking on all of our section creation logic.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@264204 91177308-0d34-0410-b5e6-96231b3b80d8