Change getReferenceInfo/getPairReferenceInfo to use new Error handling.  NFC.
Adds a GenericError class to lld/Core which can carry a string.  This is
analygous to the dynamic_error we currently use in lld/Core.

Use this GenericError instead of make_dynamic_error_code.  Also, provide
an implemention of GenericError::convertToErrorCode which for now converts
it in to the dynamic_error_code we used to have.  This will go away once
all the APIs are converted.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@264910 91177308-0d34-0410-b5e6-96231b3b80d8