Refactor R_RELAX_TLS_* value computation.
This makes it explicit that each R_RELAX_TLS_* is equivalent to some
other expression.

With this I think we are at a sweet spot for how much is done in
Target.cpp. I did experiment with moving *all* the value math out of it.
It has the advantage that we know the final value in target independent
code, but it gets quite verbose.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@270277 91177308-0d34-0410-b5e6-96231b3b80d8