Test --demangle feature.
Previously, it was not tested because the test was written in
a way that it passed on a platform that does not support
abi::__cxa_demangle. Now we restrict this test to Unix (by adding
"REQUIRES: shell") and assume that it always demangle symbols.
Thanks to Davide to find out the issue.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@274808 91177308-0d34-0410-b5e6-96231b3b80d8