Delete EhInputSection::getOffset.
We no longer need it for relocations in .eh_frame.

The only relocations that point to .eh_frame are the ones trying to
find the output .eh_frame.

This actually fixes a bug in the symbol value code. It was not
handling -1 as an indicator for a piece not being included in the
output.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@276175 91177308-0d34-0410-b5e6-96231b3b80d8