COFF: handle /debugtype option
Add the support infrastructure for the /debugtype option which takes a comma
delimited list of debug info to generate. The defaults are based on other
options potentially (/driver or /profile). This sets up the infrastructure to
allow us to emit RSDS records to get "build id" equivalents on COFF (similar to
binutils).

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@278056 91177308-0d34-0410-b5e6-96231b3b80d8