Change all the libSystem test files to be dylibs instead of normalized files.  Currently NFC.
A future commit will change when we choose to add an LC_LOAD_DYLIB to the final image.  Currently
we do this when an atom is used, but we need to do it when a dylib is referenced on the cmdline as
this matches ld64.

To allow this change, libsystem (and other future yaml files representing dylibs) need to be dylibs
so that the loader can see to add them to the referenced dylib list.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@278386 91177308-0d34-0410-b5e6-96231b3b80d8