Only restrict order if both sections are in the script.
This matches gold and bfd behavior and is required to handle some scripts.

The script has to assume where PT_LOADs start in order to align that
spot. If we don't allow section it doesn't know about to move to the
middle, we can need more PT_LOADs and those will not be aligned.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@282015 91177308-0d34-0410-b5e6-96231b3b80d8