Allow use define symbols to override linker defined ones.
I hit an internal linker script that was defining _DYNAMIC instead of
letting the linker do it. It turns out that both bfd and gold allow
that.

This is pretty easy to implement, just make the linker defined symbol
weak. This should have no impact in the case where there is no user
defined symbol: The visibility is hidden, which causes the output to
still be local.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@287260 91177308-0d34-0410-b5e6-96231b3b80d8