Add type records to TPI stream.
I don't think the data I add to a TPI stream in this patch is correct,
but at least it can be displayed using llvm-pdbdump. Until I add more
streams to a PDB file, I'm not able to know whether the data will be
accepted by MSVC tools or not.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@289183 91177308-0d34-0410-b5e6-96231b3b80d8