COFF: Support both /lldmap and /lldmap:<file>
The former option bases the filename on the output name, e.g. if the
link output is a.exe, the map will be written to a.map. This matches the
behaviour of link.exe's /MAP option and is useful for creating a map
file of each executable when building a large project.

Differential Revision: https://reviews.llvm.org/D27595

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@289271 91177308-0d34-0410-b5e6-96231b3b80d8