Port r292146.
Give priority to linker scripts over preemption.

LLD exports symbols that are also present in used shared libraries to
make sure they are preempted at runtime. That is a reasonable default,
but we must allow for it to be overwritten with linker script. If we
don't, libraries that expect to be able to hide a c++ delete operator
will fail.

This should fix the firebird build.


git-svn-id: https://llvm.org/svn/llvm-project/lld/branches/release_40@292370 91177308-0d34-0410-b5e6-96231b3b80d8