Port r291944.
Don't add DT_INIT/DT_FINI for undef and shared symbols.

The freebsd dynamic linker doesn't check if the value is null (and it
is reasonable for it to do that). That means that producing a .so with
a null DT_INIT/DT_FINI causes the base address to be called.

This should fix the libreoffice build.


git-svn-id: https://llvm.org/svn/llvm-project/lld/branches/release_40@292429 91177308-0d34-0410-b5e6-96231b3b80d8