Replace MergeOutputSection with a synthetic section.
With a synthetic merge section we can have, for example, a single
.rodata section with stings, fixed sized constants and non merge
constants.

I can be simplified further by not setting Entsize, but that is
probably better done is a followup patch.

This should allow some cleanup in the linker script code now that
every output section command maps to just one output section.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@294005 91177308-0d34-0410-b5e6-96231b3b80d8