Move SymbolTable<ELFT>::Sections out of the class.
The list of all input sections was defined in SymbolTable class for a
historical reason. The list itself is not a template. However, because
SymbolTable class is a template, we needed to pass around ELFT to access
the list. This patch moves the list out of the class so that it doesn't
need ELFT.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@296309 91177308-0d34-0410-b5e6-96231b3b80d8