Comment on the typical/simple case of VA calculation
There are many special cases and a layer of abstraction or two in the
way, but the VA calculation in the typical case is actually very simple
and probably makes perfect sense even to somebody new to linkers.

Also, this line brings together many components and is a good place to
start understanding the linker (or improve one's existing
understanding).

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@296451 91177308-0d34-0410-b5e6-96231b3b80d8