Make gc a bit more aggressive.
We were not gcing any section whose name was a C identifier. Both gold
and bfd only keep those if they are used.

To avoid having to create the __start/__stop symbols early or doing
string lookups in resolvedReloc, this patch just looks for undefined
symbols __start/__stop to decide if a section is needed or not.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@296723 91177308-0d34-0410-b5e6-96231b3b80d8