Compute Config member function return values only once.
We had a few Config member functions that returns configuration values.
For example, we had is64() which returns true if the target is 64-bit.
The return values of these functions are constant and never change.

This patch is to compute them only once to make it clear that they'll
never change.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@298168 91177308-0d34-0410-b5e6-96231b3b80d8