Fix accounting of tbss.
We were correctly computing the size contribution of a .tbss input
section (it is none), but we were incorrectly considering the
alignment of the output section: it was advancing Dot instead of
ThreadBssOffset.

As far as I can tell this was always wrong in our linkerscript
implementation, but that became more visible now that the code is
shared with the non linker script case.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@302107 91177308-0d34-0410-b5e6-96231b3b80d8