Refactor gdb index creation.
I have a patch to let DwarfContext defer to lld for getting section
contents and relocations.

That is a pretty big performance improvement.

This is just a refactoring to make that easier to do.

This change makes the *creation* of gdb index a dedicated step and
makes that templated. That is so that we can uses Elf_Rel in the code.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@307867 91177308-0d34-0410-b5e6-96231b3b80d8