ELF: Reword a somewhat misleading error message.
PT_TLS is a type of program header, so we wouldn't expect to see
one in an object file. This error should probably be referring to
the fact that we didn't see a section with the flag SHF_TLS, which
would normally cause us to create a PT_TLS program header.

Differential Revision: https://reviews.llvm.org/D35395

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@307983 91177308-0d34-0410-b5e6-96231b3b80d8