Correct ALIGN expression when inside a section.
When given

foobar = ALIGN(., 0x100);

my expectation from what the manual says is that the final address of
foobar will be aligned. It seems that bfd aligns the offset in the
section, which causes some odd results if the section is not 0x100
aligned. Gold aligns the address.

This changes lld to align the final address.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@312979 91177308-0d34-0410-b5e6-96231b3b80d8