Handle input section liveness only in MarkLive.cpp.
The condition whether a section is alive or not by default
is becoming increasingly complex, so the decision of garbage
collection is spreading over InputSection.h and MarkLive.cpp,
which is not a good state.

This moves the code to MarkLive.cpp, to keep the file the central
place to make decisions about garbage collection.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@315384 91177308-0d34-0410-b5e6-96231b3b80d8