Bring back part of r315721.
The original patch was reported to break a libcxx bot. Unfortunately
I could not reproduce the issue and the bot log is not available anymore.

This patch changes the handling of just lazy symbols. I will watch the
bot and see it breaks.

Original message:
--------------------------------------------------------------
Handle shared and lazy symbol in the gnu hash construction.

We were not considering those symbols undefined and they could end up
in the end of the dynamic symbol table.
--------------------------------------------------------------

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@316004 91177308-0d34-0410-b5e6-96231b3b80d8