[WebAssembly] Add test of indirect call to an otherwise unused type
In this case we are calling a function pointer which
a type that doesn't otherwise exist in the code.
Clearly this code can't would trap if it was ever
called (because there is not such function that
the pointer can resolve to).

But it should valid and compile and link and validation
time.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@321134 91177308-0d34-0410-b5e6-96231b3b80d8