Warn on nonexistent comdat sections in an /order file.
I didn't implement the feature in the original patch because I didn't
come up with an idea to do that easily and efficiently. Turned out that
that is actually easy to implement.

In this patch, we collect comdat sections before gc is run and warn on
nonexistent symbols in an order file.

Differential Revision: https://reviews.llvm.org/D42658

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@323699 91177308-0d34-0410-b5e6-96231b3b80d8