Refactor how we decide which sections to sort.
This is a bit more verbose, but it has a few advantages.

The logic on what to do with special sections like .init_array is not
duplicated. Before we would need keep isKnownNonreorderableSection in
sync.

I think with this the call graph based sorting can be implemented by
"just" returning a new order from buildSectionOrder.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@324744 91177308-0d34-0410-b5e6-96231b3b80d8