Improve duplicated version handling.
It looks like the problem that caused us to originally warn instead of
error was that of a symbol being assigned to the same version twice.

Now we don't warn if a symbol is assigned to the same version twice,
but error if it is assigned to multiple.

This fixes pr28342.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@326813 91177308-0d34-0410-b5e6-96231b3b80d8