[CMake] Properly conditionalize installation of lld libraries
In what appears to be a copy-and-paste error, lld currently only
installs libraries if the lld tools are configured to build.

Instead, lld should allow the libraries to be installed even if the lld
tools are not being built. Additionally, if users want to only install
the tools and not the libraries, the LLVM way of doing that is by
checking for LLVM_INSTALL_TOOLCHAIN_ONLY.

This fixes PR35960.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@327126 91177308-0d34-0410-b5e6-96231b3b80d8