Re-implement --just-symbols as a regular object file.
I tried a few different designs to find a way to implement it without
too much hassle and settled down with this. Unlike before, object files
given as arguments for --just-symbols are handled as object files, with
an exception that their section tables are handled as if they were all
null.

Differential Revision: https://reviews.llvm.org/D42025

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@328852 91177308-0d34-0410-b5e6-96231b3b80d8