Replace SharedSymbols with Defined when creating copy relocations.
This is slightly simpler to read IMHO. Now if a symbol has a position
in the file, it is Defined.

The main motivation is that with this a SharedSymbol doesn't need a
section, which reduces the size of SymbolUnion.

With this the peak allocation when linking chromium goes from 568.1 to
564.2 MB.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@330966 91177308-0d34-0410-b5e6-96231b3b80d8