[ELF] Clarify help wording for --symbol-ordering-file
The input file for this option should contain a list of symbols, not a
list of sections, so explicitly refer to ordering symbols (but keep the
reference to laying out sections, since that's how the option must
operate). Referring to the file itself as the "symbol ordering file" is
consistent with --warn-symbol-ordering and less ambiguous than "symbol
file" (albeit slightly redundant).

Differential Revision: https://reviews.llvm.org/D46099

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@331000 91177308-0d34-0410-b5e6-96231b3b80d8