[ELF][MIPS] Fill a primary-GOT as much as possible
While building a Global Offset Table try to fill the primary GOT as much
as possible because the primary GOT can be accessed in the most
effective way. If it is not possible, try to fill the last GOT in the
multi-GOT list, and finally create a new GOT if both attempts failed.

git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@335140 91177308-0d34-0410-b5e6-96231b3b80d8