lld-link: Remove /msvclto option
This was useful for LTO bringup in lld-link while lld couldn't write PDBs. Now
that it can, this should no longer be needed. Hopefully the flag is obscure
enough and recent enough, that nobody uses it – but if somebody should use it,
they should be able to just stop passing it and things should continue to work.

https://reviews.llvm.org/D50139


git-svn-id: https://llvm.org/svn/llvm-project/lld/trunk@338615 91177308-0d34-0410-b5e6-96231b3b80d8